Refactor app.py (#582)
* Refactor app.py

* Move to `create_app.py`

* Simplify tests

* Update Dockerfile
