+++ code/tests/test_app.py
@@ -2,13 +2,28 @@ import json
 import os
 import pytest
 from unittest.mock import MagicMock, Mock, patch
+from create_app import create_app
 
-from app import app
+
+@pytest.fixture
+def client():
+    return create_app().test_client()
+
+
+@pytest.fixture
+def env_helper_mock(autouse=True):
+    patcher = patch("create_app.EnvHelper")
+
+    EnvHelperMock = patcher.start()
+    mock = EnvHelperMock.return_value
+    yield mock
+
+    patcher.stop()
 
 
 class TestConfig:
-    def test_returns_correct_config(self):
-        response = app.test_client().get("/api/config")
+    def test_returns_correct_config(self, client):
+        response = client.get("/api/config")
 
         assert response.status_code == 200
         assert response.json == {
@@ -39,16 +54,16 @@ class TestConversationCustom:
             ],
         }
 
-    @patch("app.get_message_orchestrator")
+    @patch("create_app.get_message_orchestrator")
     @patch(
         "backend.batch.utilities.helpers.ConfigHelper.ConfigHelper.get_active_config_or_default"
     )
-    @patch("app.env_helper")
     def test_converstation_custom_returns_correct_response(
         self,
-        env_helper_mock,
         get_active_config_or_default_mock,
         get_message_orchestrator_mock,
+        env_helper_mock,
+        client,
     ):
         # given
         get_active_config_or_default_mock.return_value.orchestrator.return_value = (
@@ -62,7 +77,7 @@ class TestConversationCustom:
         env_helper_mock.AZURE_OPENAI_MODEL = self.openai_model
 
         # when
-        response = app.test_client().post(
+        response = client.post(
             "/api/conversation/custom",
             headers={"content-type": "application/json"},
             json=self.body,
@@ -78,10 +93,13 @@ class TestConversationCustom:
             "object": "response.object",
         }
 
-    @patch("app.get_message_orchestrator")
-    @patch("app.get_orchestrator_config")
+    @patch("create_app.get_message_orchestrator")
+    @patch("create_app.get_orchestrator_config")
     def test_converstation_custom_calls_message_orchestrator_correctly(
-        self, get_orchestrator_config_mock, get_message_orchestrator_mock
+        self,
+        get_orchestrator_config_mock,
+        get_message_orchestrator_mock,
+        client,
     ):
         # given
         get_orchestrator_config_mock.return_value = self.orchestrator_config
@@ -93,7 +111,7 @@ class TestConversationCustom:
         os.environ["AZURE_OPENAI_MODEL"] = self.openai_model
 
         # when
-        app.test_client().post(
+        client.post(
             "/api/conversation/custom",
             headers={"content-type": "application/json"},
             json=self.body,
@@ -107,15 +125,15 @@ class TestConversationCustom:
             orchestrator=self.orchestrator_config,
         )
 
-    @patch("app.get_orchestrator_config")
+    @patch("create_app.get_orchestrator_config")
     def test_converstation_custom_returns_error_resonse_on_exception(
-        self, get_orchestrator_config_mock
+        self, get_orchestrator_config_mock, client
     ):
         # given
         get_orchestrator_config_mock.side_effect = Exception("An error occurred")
 
         # when
-        response = app.test_client().post(
+        response = client.post(
             "/api/conversation/custom",
             headers={"content-type": "application/json"},
             json=self.body,
@@ -127,10 +145,10 @@ class TestConversationCustom:
             "error": "Exception in /api/conversation/custom. See log for more details."
         }
 
-    @patch("app.get_message_orchestrator")
-    @patch("app.get_orchestrator_config")
+    @patch("create_app.get_message_orchestrator")
+    @patch("create_app.get_orchestrator_config")
     def test_converstation_custom_allows_multiple_messages_from_user(
-        self, get_orchestrator_config_mock, get_message_orchestrator_mock
+        self, get_orchestrator_config_mock, get_message_orchestrator_mock, client
     ):
         """This can happen if there was an error getting a response from the assistant for the previous user message."""
 
@@ -140,8 +158,6 @@ class TestConversationCustom:
         message_orchestrator_mock = Mock()
         message_orchestrator_mock.handle_message.return_value = self.messages
         get_message_orchestrator_mock.return_value = message_orchestrator_mock
-
-        os.environ["AZURE_OPENAI_MODEL"] = self.openai_model
 
         body = {
             "conversation_id": "123",
@@ -157,7 +173,7 @@ class TestConversationCustom:
         }
 
         # when
-        response = app.test_client().post(
+        response = client.post(
             "/api/conversation/custom",
             headers={"content-type": "application/json"},
             json=body,
@@ -275,34 +291,27 @@ class TestConversationAzureByod:
         self.stop_sequence = "\n|STOP"
 
     @pytest.fixture(autouse=True)
-    def env_helper_mock(self):
-        patcher = patch("app.env_helper")
-        mock = patcher.start()
-
+    def setup_env_helper_mock(self, env_helper_mock):
         # These are the default values for the env_helper
         # They can be overridden within each test
-        mock.AZURE_OPENAI_SYSTEM_MESSAGE = self.system_message
-        mock.AZURE_OPENAI_MODEL = self.openai_model
-        mock.AZURE_OPENAI_ENDPOINT = self.openai_endpoint
-        mock.AZURE_OPENAI_API_VERSION = self.openai_api_version
-        mock.AZURE_OPENAI_API_KEY = self.openai_api_key
-        mock.AZURE_SEARCH_KEY = self.search_key
-        mock.AZURE_TOKEN_PROVIDER.return_value = self.token
-        mock.AZURE_OPENAI_TEMPERATURE = self.temperature
-        mock.AZURE_OPENAI_MAX_TOKENS = self.max_tokens
-        mock.AZURE_OPENAI_TOP_P = self.top_p
-        mock.AZURE_OPENAI_STOP_SEQUENCE = self.stop_sequence
-        mock.SHOULD_STREAM = True
-        mock.AZURE_AUTH_TYPE = "keys"
-        mock.should_use_data.return_value = True
+        env_helper_mock.AZURE_OPENAI_SYSTEM_MESSAGE = self.system_message
+        env_helper_mock.AZURE_OPENAI_MODEL = self.openai_model
+        env_helper_mock.AZURE_OPENAI_ENDPOINT = self.openai_endpoint
+        env_helper_mock.AZURE_OPENAI_API_VERSION = self.openai_api_version
+        env_helper_mock.AZURE_OPENAI_API_KEY = self.openai_api_key
+        env_helper_mock.AZURE_SEARCH_KEY = self.search_key
+        env_helper_mock.AZURE_TOKEN_PROVIDER.return_value = self.token
+        env_helper_mock.AZURE_OPENAI_TEMPERATURE = self.temperature
+        env_helper_mock.AZURE_OPENAI_MAX_TOKENS = self.max_tokens
+        env_helper_mock.AZURE_OPENAI_TOP_P = self.top_p
+        env_helper_mock.AZURE_OPENAI_STOP_SEQUENCE = self.stop_sequence
+        env_helper_mock.SHOULD_STREAM = True
+        env_helper_mock.AZURE_AUTH_TYPE = "keys"
+        env_helper_mock.should_use_data.return_value = True
 
-        yield mock
-
-        patcher.stop()
-
-    @patch("app.requests.Session")
+    @patch("create_app.requests.Session")
     def test_converstation_azure_byod_returns_correct_response_when_streaming_with_data_keys(
-        self, get_requests_session_mock
+        self, get_requests_session_mock, client
     ):
         # given
         mock_session = get_requests_session_mock.return_value
@@ -310,7 +319,7 @@ class TestConversationAzureByod:
         mock_session.post = Mock(return_value=response_mock)
 
         # when
-        response = app.test_client().post(
+        response = client.post(
             "/api/conversation/azure_byod",
             headers={"content-type": "application/json"},
             json=self.body,
@@ -338,9 +347,9 @@ class TestConversationAzureByod:
         }
         assert request_headers["api-key"] == self.openai_api_key
 
-    @patch("app.requests.Session")
+    @patch("create_app.requests.Session")
     def test_converstation_azure_byod_returns_correct_response_when_streaming_with_data_rbac(
-        self, get_requests_session_mock, env_helper_mock
+        self, get_requests_session_mock, env_helper_mock, client
     ):
         # given
         mock_session = get_requests_session_mock.return_value
@@ -349,7 +358,7 @@ class TestConversationAzureByod:
         env_helper_mock.AZURE_AUTH_TYPE = "rbac"
 
         # when
-        response = app.test_client().post(
+        response = client.post(
             "/api/conversation/azure_byod",
             headers={"content-type": "application/json"},
             json=self.body,
@@ -376,9 +385,9 @@ class TestConversationAzureByod:
         }
         assert request_headers["Authorization"] == f"Bearer {self.token}"
 
-    @patch("app.requests.post")
+    @patch("create_app.requests.post")
     def test_converstation_azure_byod_returns_correct_response_when_not_streaming_with_data(
-        self, post_mock, env_helper_mock
+        self, post_mock, env_helper_mock, client
     ):
         # given
         env_helper_mock.SHOULD_STREAM = False
@@ -400,7 +409,7 @@ class TestConversationAzureByod:
         }
 
         # when
-        response = app.test_client().post(
+        response = client.post(
             "/api/conversation/azure_byod",
             headers={"content-type": "application/json"},
             json=self.body,
@@ -431,9 +440,9 @@ class TestConversationAzureByod:
             ],
         }
 
-    @patch("app.requests.Session")
+    @patch("create_app.requests.Session")
     def test_converstation_azure_byod_receives_error_from_search_when_streaming_with_data(
-        self, get_requests_session_mock
+        self, get_requests_session_mock, client
     ):
         # given
         mock_session = get_requests_session_mock.return_value
@@ -441,7 +450,7 @@ class TestConversationAzureByod:
         mock_session.post = Mock(return_value=response_mock)
 
         # when
-        response = app.test_client().post(
+        response = client.post(
             "/api/conversation/azure_byod",
             headers={"content-type": "application/json"},
             json=self.body,
@@ -451,16 +460,16 @@ class TestConversationAzureByod:
         assert response.status_code == 200
         assert b'"error": "An error occurred\\n"' in response.data
 
-    @patch("app.requests.Session")
+    @patch("create_app.requests.Session")
     def test_converstation_azure_byod_throws_exception_when_streaming_with_data(
-        self, get_requests_session_mock
+        self, get_requests_session_mock, client
     ):
         # given
         mock_session = get_requests_session_mock.return_value
         mock_session.post.side_effect = ValueError("Test exception")
 
         # when
-        response = app.test_client().post(
+        response = client.post(
             "/api/conversation/azure_byod",
             headers={"content-type": "application/json"},
             json=self.body,
@@ -470,15 +479,15 @@ class TestConversationAzureByod:
         assert response.status_code == 200
         assert b'{"error": "Test exception"}\n' in response.data
 
-    @patch("app.conversation_with_data")
+    @patch("create_app.conversation_with_data")
     def test_converstation_azure_byod_returns_500_when_exception_occurs(
-        self, conversation_with_data_mock
+        self, conversation_with_data_mock, client
     ):
         # given
         conversation_with_data_mock.side_effect = Exception("Test exception")
 
         # when
-        response = app.test_client().post(
+        response = client.post(
             "/api/conversation/azure_byod",
             headers={"content-type": "application/json"},
             json=self.body,
@@ -490,9 +499,9 @@ class TestConversationAzureByod:
             "error": "Exception in /api/conversation/azure_byod. See log for more details."
         }
 
-    @patch("app.AzureOpenAI")
+    @patch("create_app.AzureOpenAI")
     def test_converstation_azure_byod_returns_correct_response_when_not_streaming_without_data_keys(
-        self, azure_openai_mock, env_helper_mock
+        self, azure_openai_mock, env_helper_mock, client
     ):
         # given
         env_helper_mock.should_use_data.return_value = False
@@ -511,7 +520,7 @@ class TestConversationAzureByod:
         openai_client_mock.chat.completions.create.return_value = openai_create_mock
 
         # when
-        response = app.test_client().post(
+        response = client.post(
             "/api/conversation/azure_byod",
             headers={"content-type": "application/json"},
             json=self.body,
@@ -553,9 +562,9 @@ class TestConversationAzureByod:
             stream=False,
         )
 
-    @patch("app.AzureOpenAI")
+    @patch("create_app.AzureOpenAI")
     def test_converstation_azure_byod_returns_correct_response_when_not_streaming_without_data_rbac(
-        self, azure_openai_mock, env_helper_mock
+        self, azure_openai_mock, env_helper_mock, client
     ):
         # given
         env_helper_mock.should_use_data.return_value = False
@@ -576,7 +585,7 @@ class TestConversationAzureByod:
         openai_client_mock.chat.completions.create.return_value = openai_create_mock
 
         # when
-        response = app.test_client().post(
+        response = client.post(
             "/api/conversation/azure_byod",
             headers={"content-type": "application/json"},
             json=self.body,
@@ -618,9 +627,9 @@ class TestConversationAzureByod:
             stream=False,
         )
 
-    @patch("app.AzureOpenAI")
+    @patch("create_app.AzureOpenAI")
     def test_converstation_azure_byod_returns_correct_response_when_streaming_without_data(
-        self, azure_openai_mock, env_helper_mock
+        self, azure_openai_mock, env_helper_mock, client
     ):
         # given
         env_helper_mock.should_use_data.return_value = False
@@ -639,7 +648,7 @@ class TestConversationAzureByod:
         openai_client_mock.chat.completions.create.return_value = [mock_response]
 
         # when
-        response = app.test_client().post(
+        response = client.post(
             "/api/conversation/azure_byod",
             headers={"content-type": "application/json"},
             json=self.body,
